+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 * Updated to `http-interop/http-middleware#0.3`
 
+### Added
+
+* New option `methods()` to configure the allowed methods
+
 ## 0.1.0 - 2016-10-04
 
 First version